requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
alembic
argon2-cffi
python-jose[cryptography]